[#174] refactor: 이전 검색내역 style 코드 분리
