+++ src/components/common/Header/style.ts
@@ -1,30 +1,7 @@
-import { motion } from 'framer-motion';
 import styled from 'styled-components';
 
 import * as Layout from '@/components/common/layout';
 import * as Font from '@/styles/font';
-
-export const Button = styled(motion.button)`
-  outline: none;
-  border: none;
-  background: none;
-  width: 30px;
-  height: 30px;
-  border-radius: 8px;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  padding: 0;
-  left: 0;
-
-  :hover {
-    background-color: ${props => props.theme.colors.gray[0]};
-    transition: all 500ms;
-  }
-`;
 
 export const HeaderWrapper = styled.header`
   display: flex;
@@ -50,64 +27,12 @@ export const StyleTextLogo = styled(Font.Medium)`
   text-align: center;
 `;
 
-export const SearchHeader = styled(Layout.HStack)`
-  padding: 5vh calc(1vw + 16px) 14px 60px;
-  height: 42px;
-  min-height: 42px;
-
-  @media (padding-top: 5svh) {
-    padding: 5svh calc(1vw + 16px) 2svh 60px;
-  }
-
-  @media (width>= 768px) {
-    padding: 5vh calc(1vw + 16px) 16px 60px;
-  }
-`;
-
 export const Line = styled.hr`
   height: 1px;
   width: 90%;
   margin: 0;
   border: none;
   background: ${props => props.theme.colors.gray[1]};
-`;
-
-export const ListBox = styled(motion.div)`
-  display: flex;
-  justify-content: flex-start;
-  flex-direction: column;
-
-  width: 80%;
-  height: 100vh;
-  /* for iphone safari 100vh bug fix */
-  height: -webkit-fill-available;
-  height: fill-available;
-
-  z-index: 2;
-  background: #ffffff;
-`;
-
-export const Absolute = styled(Layout.VStack)`
-  position: absolute;
-  width: 375px;
-  @media (max-width: 768px) {
-    width: 100%;
-  }
-`;
-
-export const Background = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  backdrop-filter: blur(2px);
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 1;
-`;
-
-export const Href = styled.div`
-  cursor: pointer;
 `;
 
 export const UserFormHref = styled(Layout.VStack)`
@@ -117,39 +42,4 @@ export const UserFormHref = styled(Layout.VStack)`
 export const TagRoundOrang = styled(Layout.Box)`
   border-radius: 24px;
   border: 1px solid ${({ theme }) => theme.colors.mainColor};
-`;
-
-export const ResponsivePosition = styled(Layout.VStack)`
-  padding: 0 50px;
-
-  @media (width >= 768px) {
-    padding: 0 calc(1vh + 16px + 30px);
-  }
-`;
-
-export const ToastBackground = styled.div`
-  z-index: 2;
-  div {
-    background-color: rgba(0, 0, 0, 0.8);
-  }
-`;
-
-export const YScroll = styled(Layout.VStack)`
-  overflow: auto;
-  ::-webkit-scrollbar,
-  ::-webkit-scrollbar-thumb {
-    width: 14px;
-    border-radius: 7px;
-    background-clip: padding-box;
-    border: 5px solid white;
-    margin-right: 5px;
-  }
-
-  ::-webkit-scrollbar-track-piece {
-    width: 5px;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    background: ${({ theme }) => theme.colors.gray[2]};
-  }
 `;
